requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
peft>=0.7.0
datasets>=2.16.0
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.65.0
accelerate>=0.25.0
evaluate>=0.4.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.13.0
